requirements.txt
africastalking==1.2.7
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contextlib2==21.6.0
cryptography==42.0.4
Django==5.0.2
django-cors-middleware==1.5.0
django-extensions==3.2.3
django-filter==23.5
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
idna==3.6
inflection==0.5.1
jwcrypto==1.5.4
oauthlib==3.2.2
pycparser==2.21
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
schema==0.7.5
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
